LayoutTests/imported/w3c/web-platform-tests/IndexedDB/file_support.sub-expected.txt



FAIL Saves and loads back File objects from IndexedDB assert_equals: expected 1 but got 0
